working on data harmon

+++ UMPG.USL.API.Data/DataHarmonization/SnapshotConfigurationRepository.cs
@@ -32,7 +32,11 @@ namespace UMPG.USL.API.Data.DataHarmonization
             using (var context = new AuthContext())
             {
                 var licenseProduct =
-                    context.Snapshot_Configurations.First(_ => _.SnapshotConfigId == snapshotLicenseProductId);
+                    context.Snapshot_Configurations.FirstOrDefault(_ => _.SnapshotConfigId == snapshotLicenseProductId);
+                if (licenseProduct == null)
+                {
+                    return false;
+                }
                 context.Snapshot_Configurations.Attach(licenseProduct);
                 context.Snapshot_Configurations.Remove(licenseProduct);
                 try
